ltl/requirements.txt
absl-py==0.8.1
astor==0.7.1
-e git+https://github.com/openai/baselines.git@665b888eeb688396894455a0d94febc4f712e0c0#egg=baselines
certifi==2019.9.11
cffi==1.13.1
Click==7.0
cloudpickle==1.2.2
decorator==4.4.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.7
graphviz==0.13
grpcio==1.23.0
gym==0.15.3
h5py==2.10.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
networkx==2.3
nltk==3.4.5
numpy==1.17.3
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==0+untagged.49.gdbede45.dirty
Pillow==6.2.1
protobuf==3.10.0
pycparser==2.19
pyglet==1.3.2
scipy==1.3.1
six==1.12.0
tensorboard==2.0.0
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
torch==1.3.0
torchvision==0.4.1
tqdm==4.37.0
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
